Added flask sessions to gitignore
