+++ web-app/templates/index.html
@@ -6,7 +6,6 @@
         <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500&display=swap" rel="stylesheet">
         <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
         <link rel="stylesheet", type="text/css", href="../static/style.css">
-        <script src="../static/index.js"></script>
     </head>
 
     <body>
